+++ nixos/modules/programs/nix.nix
@@ -4,13 +4,15 @@
 ## * Keep derivations and build outputs (good for developers)                 ##
 ## * Use all the cores to build deviravions                                   ##
 ## * Automatically delete generations older than 30 days every day at 21:00   ##
+## * Optionally enable `nh` for both system operations and garbage collection ##
 ##                                                                            ##
 ################################################################################
 
 { config, lib, pkgs, ... }:
 
 let
-  inherit (lib) mkDefault mkEnableOption mkIf;
+  inherit (lib) mkDefault mkEnableOption mkIf mkOption;
+  inherit (lib.types) bool;
   inherit (pkgs) stdenv;
 
   cfg = config.confkit.programs.nix;
@@ -19,6 +21,19 @@ in
 {
   options.confkit.programs.nix = {
     enable = mkEnableOption "the confkit configuration for Nix";
+
+    nh.enable = mkOption {
+      type = bool;
+      default = false;
+      example = true;
+      description = ''
+        Whether to enable nh.
+
+        If nh is enabled, the system flake is automatically selected and
+        `nh clean all` is used for garbage collection instead of
+        `nix-collect-garbage`.
+      '';
+    };
   };
 
   config = mkIf cfg.enable {
@@ -34,7 +49,7 @@ in
         keep-outputs = mkDefault true;
       };
 
-      gc =
+      gc = mkIf (!cfg.nh.enable) (
         let
           gc-common = {
             automatic = mkDefault true;
@@ -45,7 +60,20 @@ in
           interval = mkDefault { Hour = 21; Minute = 0; };
         } else gc-common // {
           dates = mkDefault "21:00";
-        };
+        }
+      );
+    };
+
+    programs.nh = mkIf cfg.nh.enable {
+      enable = true;
+      flake = mkDefault "/config/Nix/${config.networking.hostName}";
+
+      # Garbage-collect at the end of each month.
+      clean = {
+        enable = mkDefault true;
+        dates = mkDefault "21:00";
+        extraArgs = mkDefault "--nogcroots --keep-since 1M";
+      };
     };
   };
 }